Neo can now identify its location
Added an entityData class that can be attached to every player in order
for Neo to track their status and location. Neo answers the question 'where are you' by checking its own status in the DB. The code is kind of dirty since I'm
trying to get the unit tests to pass as quickly as possible. I will clean
up later as I add more exception handling.

+++ Assets/NEO/test/Editor/NeoMemoryTest.cs
@@ -4,16 +4,20 @@ using UnityEngine.TestTools;
 using NUnit.Framework;
 using System.Collections;
 using System.Collections.Generic;
+using Mono.Data.Sqlite;
+using System.Data;
 
 public class NeoMemoryTest {
 
 	private NeoMemory neoMemory;
 	private ObjectData apple;
+	private string TEST_CONNECTION_STRING;
 
 	[SetUp]
 	public void SetupNeoMemory(){
 		neoMemory = NeoMemory.getInstance ();
-		neoMemory.setconnectionString("URI=file:" + Application.dataPath + "/test/neo_brain_test.db");
+		TEST_CONNECTION_STRING = "URI=file:" + Application.dataPath + "/test/neo_brain_test.db";
+		neoMemory.setconnectionString(TEST_CONNECTION_STRING);
 		apple = new ObjectData("apple", "red", new List<string>(new string[]{"fruit", "food"}), 200, 200);
 	}
 
@@ -24,11 +28,28 @@ public class NeoMemoryTest {
 		Assert.AreEqual(neoMemory.getObjects(), new List<string>(new string[]{"apple"}));
 	}
 
+	[Test]
+	public void NeoMemoryUpdatesPersonLocation() {
+		LocationData data = new LocationData ("kitchen", 200, 200);
+		neoMemory.MemorizeLocation (data);
+		neoMemory.UpdatePersonLocation ("you", "kitchen");
+		using (IDbConnection dbConnection = new SqliteConnection(TEST_CONNECTION_STRING)) {
+			dbConnection.Open ();
+
+			using (IDbCommand dbCommand = dbConnection.CreateCommand ()) {
+				Assert.AreEqual(1, DBUtils.getId(DBUtils.PERSON_LOCATION, DBUtils.PEOPLE_TABLE, DBUtils.PEOPLE_COL, "you", dbCommand));
+			}
+			dbConnection.Close ();
+		}
+	}
+
+
+
 	[TearDown]
 	public void cleanUpDB(){
 		List<string> tables = neoMemory.getTableNames();
 		foreach (string table in tables){
-			if (table != DBUtils.ATTRIBUTES_TABLE) {
+			if (table.ToUpper() != DBUtils.ATTRIBUTES_TABLE && table.ToUpper() != DBUtils.PEOPLE_TABLE) {
 				neoMemory.runCommand ("DELETE FROM " + table);
 			}
 		}
